[IMP] fieldservice_stock: Split equipment
Remove the equipment related functionality from the fieldservice_stock module and make into separate module
