+++ fieldservice_equipment_stock/README.rst
@@ -1,0 +1,137 @@
+===============================
+Field Service - Stock Equipment
+===============================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Ffield--service-lightgray.png?logo=github
+    :target: https://github.com/OCA/field-service/tree/12.0/fieldservice_stock
+    :alt: OCA/field-service
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/field-service-12-0/field-service-12-0-fieldservice_stock
+    :alt: Translate me on Weblate
+.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
+    :target: https://runbot.odoo-community.org/runbot/264/12.0
+    :alt: Try me on Runbot
+
+|badge1| |badge2| |badge3| |badge4| |badge5| 
+
+This module is an add-on for the Field Service application in Odoo.
+It allows you to track your FSM Equipments in stock moves.
+
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Configuration
+=============
+
+Products can be automatically converted into FSM Equipments.
+This is only available only for products tracked by serial number.
+This needs to be enabled both on Operation Types and Products.
+For example, we may want to create the FSM Equipment on Delivery,
+or on Receipts.
+
+To enable on Products:
+
+* Go to Inventory > Master Data > Products
+* Open the Product form, Inventory tab
+* On the "Traceability" section, make sure
+  "Tracking" is set to "By Unique Serial Number"
+* Enable the "Creates FSM Equipment" checkbox
+
+To enable on Operation Types:
+
+* Go to Inventory > Configuration > Operation Types
+* Select the intended Operation Type ("Receipts" for example)
+* On the "Traceability" section, enable the
+  "Create FSM Equipment" checkbox
+
+Usage
+=====
+
+To use this module, you need to:
+
+Known issues / Roadmap
+======================
+
+The roadmap of the Field Service application is documented on
+`Github <https://github.com/OCA/field-service/issues/1>`_.
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/field-service/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/field-service/issues/new?body=module:%20fieldservice_stock%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* Open Source Integrators
+* Brian McMaster
+
+Contributors
+~~~~~~~~~~~~
+
+* Brian McMaster <brian@mcmpest.com>
+* Sandip Mangukiya <smangukiya@opensourceintegrators.com>
+* Serpent Consulting Services Pvt. Ltd. <support@serpentcs.com>
+
+Other credits
+~~~~~~~~~~~~~
+
+The development of this module has been financially supported by:
+
+* Open Source Integrators <https://opensourceintegrators.com>
+
+Maintainers
+~~~~~~~~~~~
+
+This module is maintained by the OCA.
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+.. |maintainer-brian10048| image:: https://github.com/brian10048.png?size=40px
+    :target: https://github.com/brian10048
+    :alt: brian10048
+.. |maintainer-wolfhall| image:: https://github.com/wolfhall.png?size=40px
+    :target: https://github.com/wolfhall
+    :alt: wolfhall
+.. |maintainer-max3903| image:: https://github.com/max3903.png?size=40px
+    :target: https://github.com/max3903
+    :alt: max3903
+.. |maintainer-smangukiya| image:: https://github.com/smangukiya.png?size=40px
+    :target: https://github.com/smangukiya
+    :alt: smangukiya
+
+Current `maintainers <https://odoo-community.org/page/maintainer-role>`__:
+
+|maintainer-brian10048| |maintainer-wolfhall| |maintainer-max3903| |maintainer-smangukiya| 
+
+This module is part of the `OCA/field-service <https://github.com/OCA/field-service/tree/12.0/fieldservice_stock_equipment>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
