+++ fieldservice_equipment_stock/__manifest__.py
@@ -1,26 +1,24 @@
-# Copyright (C) 2018 - TODAY, Brian McMaster
+# Copyright (C) 2020, Brian McMaster
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
-    'name': 'Field Service - Stock',
-    'summary': 'Integrate the logistics operations with Field Service',
-    'version': '12.0.2.3.0',
+    'name': 'Field Service - Stock Equipment',
+    'summary': 'Integrate stock operations with your field service equipments',
+    'version': '12.0.1.0.0',
     'category': 'Field Service',
     'author': "Open Source Integrators, "
               "Brian McMaster, "
               "Odoo Community Association (OCA)",
     'website': 'https://github.com/OCA/field-service',
     'depends': [
-        'fieldservice',
-        'stock',
+        'fieldservice_stock',
     ],
     'data': [
         'security/ir.model.access.csv',
-        'data/fsm_stock_data.xml',
-        'views/fsm_territory.xml',
-        'views/fsm_location.xml',
-        'views/fsm_order.xml',
-        'views/stock.xml',
+        'views/fsm_equipment.xml',
+        'views/product_template.xml',
+        'views/stock_picking_type.xml',
+        'views/stock_production_lot.xml',
     ],
     'license': 'AGPL-3',
     'development_status': 'Beta',
