+++ fieldservice_equipment_stock/models/__init__.py
@@ -1,0 +1,7 @@
+from . import (
+    stock_move,
+    stock_picking_type,
+    fsm_equipment,
+    product_template,
+    stock_production_lot,
+)
