+++ fieldservice_equipment_stock/models/fsm_equipment.py
