+++ fieldservice_equipment_stock/models/product_template.py
