+++ fieldservice_equipment_stock/models/stock_picking_type.py
