+++ fieldservice_equipment_stock/models/stock_production_lot.py
