+++ fieldservice_equipment_stock/readme/CONTRIBUTORS.rst
@@ -1,0 +1,4 @@
+* Brian McMaster <brian@mcmpest.com>
+* Sandip Mangukiya <smangukiya@opensourceintegrators.com>
+* Serpent Consulting Services Pvt. Ltd. <support@serpentcs.com>
+* Marcel Savegnago <marcel.savegnago@escodoo.com.br>
