+++ fieldservice_equipment_stock/readme/USAGE.rst
@@ -1,0 +1,2 @@
+A completed stock move for a properly configured Product on a configured
+stock Operation Type will automatically create a field service equipment.
