+++ fieldservice_equipment_stock/tests/test_fsm_equipment.py
