+++ fieldservice_equipment_stock/views/fsm_equipment.xml
