+++ fieldservice_equipment_stock/views/product_template.xml
