+++ fieldservice_equipment_stock/views/stock_picking_type.xml
