+++ fieldservice_equipment_stock/views/stock_production_lot.xml
