+++ fieldservice_repair/__manifest__.py
@@ -4,12 +4,12 @@
 {
     'name': 'Field Service - Repair',
     'summary': 'Integrate Field Service orders with MRP repair orders',
-    'version': '12.0.1.1.0',
+    'version': '12.0.1.2.0',
     'category': 'Field Service',
     'author': 'Open Source Integrators, Odoo Community Association (OCA)',
     'website': 'https://github.com/OCA/field-service',
     'depends': [
-        'fieldservice_stock',
+        'fieldservice_equipment_stock',
         'repair',
     ],
     'data': [
