+++ fieldservice_stock/models/__init__.py
@@ -9,9 +9,5 @@ from . import (
     procurement_group,
     stock_rule,
     stock_picking,
-    stock_picking_type,
-    fsm_equipment,
-    product_template,
-    stock_production_lot,
     fsm_wizard,
 )
