+++ fieldservice_stock/security/ir.model.access.csv
@@ -1,6 +1,5 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
 access_stock_move_portal,access.stock.move.portal,stock.model_stock_move,base.group_portal,1,1,0,0
-access_stock_picking_type_portal,access.stock.picking.type.portal,stock.model_stock_picking_type,base.group_portal,1,0,0,0
 access_product_product_portal,access.product.product.portal,product.model_product_product,base.group_portal,1,0,0,0
 access_product_template_portal,access.product.template.portal,product.model_product_template,base.group_portal,1,0,0,0
 access_uom_uom_portal,access.uom.uom.portal,uom.model_uom_uom,base.group_portal,1,0,0,0
